Clarify Godot requirements
At the moment, the most recent commit to the main branch was created
before Godot 4 was released [1] [2]. When that commit was created, it
was accurate to tell users to download the latest version of Godot
because the latest version of Godot was a version of Godot 3. Now, that
statement is no longer accurate because the latest version of Godot is
a version of Godot 4, and Oh My Git! hasn’t been ported to Godot 4 yet.

It’s possible that we can just import Oh My Git! into Godot 4 and have
it work fine, but that would require significant testing that I’m not
going to do at the moment.

[1]: 853b8e2 (Merge pull request #144 from n4mr3g/patch-1, 2022-07-19)
[2]: <https://godotengine.org/article/godot-4-0-sets-sail/>